% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{addition}
\alias{addition}
\title{Using regression to find the network matrix}
\usage{
addition(data, lambda)
}
\arguments{
\item{data}{Matrix representing the n observations for p variable}

\item{lambda}{Tuning parameter for lasso}
}
\value{
Matrix representing the network
}
\description{
Using regression to find the network matrix
}
