% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm.iters.r
\name{admm.iters}
\alias{admm.iters}
\title{ADMM for FGL:}
\usage{
admm.iters(
  Y,
  lambda1,
  lambda2,
  penalty = "fused",
  rho = 1,
  rho.increment = 1,
  weights,
  penalize.diagonal,
  maxiter = 1000,
  tol = 1e-05,
  warm = NULL
)
}
\arguments{
\item{warm}{}
}
\value{
a list
}
\description{
ADMM for FGL:
}
