% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heter.R
\name{alikehood}
\alias{alikehood}
\title{Likelihood for alpha}
\usage{
alikehood(a, covariates, tauhat)
}
\arguments{
\item{a}{Coefficient vector for covariates}

\item{covariates}{Covariates that affect auto correlations}

\item{tauhat}{Given vector of auto correlations}
}
\value{
Likihood value for a
}
\description{
Likelihood for alpha
}
