% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{bicfunction}
\alias{bicfunction}
\title{BIC function for i.i.d normal data}
\usage{
bicfunction(data, G, Tem)
}
\arguments{
\item{data}{The data of interest}

\item{G}{The estimated network}
}
\value{
The value of bic
}
\description{
BIC function for i.i.d normal data
}
