% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{cenfunction}
\alias{cenfunction}
\title{Data censor}
\usage{
cenfunction(data, zirate = c(0.2, 0))
}
\arguments{
\item{data}{Data matrix to be censored}

\item{zirate}{Censor parameter, in which
zirate[1] is the lower percentage of data to be censored, zirate[2] is
the censor probability for the selected ones.}
}
\value{
A censored matrix
}
\description{
Data censor
}
