% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{covaheterlongraph}
\alias{covaheterlongraph}
\title{Estimates of correlation parameters and precision matrix}
\usage{
covaheterlongraph(data, covariates, rho, type, tole, lower, upper)
}
\arguments{
\item{data}{Data matrix  in which the first column is subject id, the second column is
the time points of observation.  Columns 2 to (p+2) is the observations for p variables.}

\item{rho}{Tuning parameter used in graphical lasso}

\item{type}{Type of correlation function, which can take either  "abs" or "sqr".}

\item{tole}{Error tolerance for determination of convergence of EM algorithm}

\item{lower}{Lower bound for prediction of correlation parameter tau}

\item{upper}{Upper bound for prediction of correlation parameter tau}

\item{x}{Data frame representing the covariates affecting the auto correlations. The frist column is the ids}
}
\value{
S list with three components which are the final estimate of alpha, tau and precision matrix omega
}
\description{
Estimates of correlation parameters and precision matrix
}
\author{
Jie Zhou
}
