% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{cv.lglasso}
\alias{cv.lglasso}
\title{Title}
\usage{
cv.lglasso(
  type = c("general", "expFixed", "twoPara"),
  data,
  group = NULL,
  lambda = NULL,
  nlam = 10,
  lam.min.ratio = 0.01,
  K,
  expFix = 1,
  trace = FALSE
)
}
\arguments{
\item{data}{data frame}

\item{K}{number of cross validation}

\item{B}{given network}
}
\value{
list
}
\description{
Title
}
