% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{cvError}
\alias{cvError}
\title{Title}
\usage{
cvError(data.train, data.valid, B)
}
\arguments{
\item{data.train}{training data set}

\item{data.valid}{validation dat set}

\item{B}{network matrix}
}
\value{
a scalar
}
\description{
Title
}
