% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{CVlglasso}
\alias{CVlglasso}
\title{Cross validation for \code{lglasso}}
\usage{
CVlglasso(
  data,
  K,
  group = NULL,
  random = FALSE,
  lambda = NULL,
  nlam = 10,
  lam.min.ratio = 0.01,
  expFix = 1,
  trace = FALSE
)
}
\arguments{
\item{data}{same as in \code{lglasso}}

\item{K}{fold of the cross validation}

\item{group}{same as in \code{lglasso}}

\item{random}{same as in \code{lglasso}}

\item{lambda}{tuning parameter. For one-stage model, lambda is a vector. For two-stage model,
lambda is a matrix. The first entry is to control the sparsity, the second entry is to control the
similarity of two networks.}

\item{nlam}{If \code{lambda} is NULL, then \code{nlam} set the number of tuning parameter}

\item{lam.min.ratio}{ratio of largest lambda vs smallest lambda}

\item{expFix}{same as on \code{lglasso}}

\item{trace}{whether show the process}
}
\value{
list of which the first component is the cross validation errors and the second component is the corresponding
tuning parameters
}
\description{
The function computes the cross validation errors  in \code{lglasso}.
}
