% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{cvNetwork}
\alias{cvNetwork}
\title{Title}
\usage{
cvNetwork(
  type = c("general", "expFixed", "twoPara"),
  group = NULL,
  data,
  lambda,
  K,
  expFix = 1
)
}
\arguments{
\item{data}{data frame}

\item{K}{number of cross validation}

\item{B}{given network}
}
\value{
list
}
\description{
Title
}
