% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{cvp.lglasso}
\alias{cvp.lglasso}
\title{Title}
\usage{
cvp.lglasso(
  type = c("general", "expFixed", "twoPara"),
  data,
  group = NULL,
  lambda = NULL,
  nlam = 10,
  lam.min.ratio = 0.01,
  K,
  expFix = 1,
  trace = FALSE,
  cores = 1
)
}
\arguments{
\item{cores}{}
}
\description{
Title
}
