% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_functions.R
\name{homolongraph}
\alias{homolongraph}
\title{Estiamte of precision matrix and autocorrelaton parameter for homogeneous model}
\usage{
homolongraph(data, rho, type, tole, lower, upper)
}
\arguments{
\item{data}{Data matrix  in which the first column is subject id, the second column is
the time points of observation.  Columns 2 to (p+2) is the observations for p variables.}

\item{rho}{Tuning parameter for graphical lasso}

\item{type}{Type of correlation function, which can take either  "abs" or "qua".}

\item{tole}{Error tolerance for determination of convergence of EM algorithm}

\item{lower}{Lower bound for prediction of correlation parameter tau}

\item{upper}{Upper bound for prediction of correlation parameter tau}
}
\value{
A list for estimates of precision matrix and correlation parameter for given tuning parameter
}
\description{
Estiamte of precision matrix and autocorrelaton parameter for homogeneous model
}
\author{
Jie Zhou
}
