% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{importanceEstimates}
\alias{importanceEstimates}
\title{Function for computing estimates in EM algorithm}
\usage{
importanceEstimates(importancesSample, datai, groupi, expFix = 1)
}
\arguments{
\item{importancesSample}{the random samples}

\item{datai}{the data for subject i}

\item{groupi}{specify how datai is grouped}

\item{expFix}{a scalar specifying the form of the correlation function}
}
\value{
a list of estimated
}
\description{
Function for computing estimates in EM algorithm
}
