% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{importanceSample}
\alias{importanceSample}
\title{Function for generating the samples from posterior distribution in EM algorithm}
\usage{
importanceSample(n, datai, wi, alpha, groupi, expFix = 1)
}
\arguments{
\item{n}{the number of random samples}

\item{datai}{the data for subject i}

\item{wi}{the given precision matrix}

\item{alpha}{the exponential distribution with rate alpha}

\item{groupi}{specify how datai is grouped}
}
\value{
a data frame for samples and their weights
}
\description{
Function for generating the samples from posterior distribution in EM algorithm
}
