\name{Component function for maximum likelihood function}
\alias{iss}
\title{Construct the covariance matrix}
\description{
 This function is used to construct the covariance matrix for ith subject.
}
\usage{
iss(idata,itau,ty)
}
\arguments{
  \item{idata}{The data set for ith subject.}
  \item{itau}{Correlation parameter for ith subject.}
  \item{ty}{Type of correlation function.}
}
\value{
A square matrix representing the covariance matrix
}


\author{Jie Zhou}

