% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JGL.r
\name{JGL}
\alias{JGL}
\title{admm algorithm of joint graphical lasso}
\usage{
JGL(
  Y,
  penalty = "fused",
  lambda1,
  lambda2,
  rho = 1,
  weights = "equal",
  penalize.diagonal = FALSE,
  maxiter = 500,
  tol = 1e-05,
  warm = NULL,
  return.whole.theta = FALSE,
  screening = "fast",
  truncate = 1e-05
)
}
\value{
a network list with length equal to that of Y
}
\description{
admm algorithm of joint graphical lasso
}
