% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{lglasso}
\alias{lglasso}
\title{Longitudinal graphical lasso}
\usage{
lglasso(
  data,
  lambda,
  group = NULL,
  random = FALSE,
  expFix = 1,
  N = 100,
  maxit = 50,
  tol = 10^(-2),
  lower = c(0.01, 0.01),
  upper = c(10, 10),
  w.init = NULL,
  wi.init = NULL,
  trace = FALSE,
  ...
)
}
\arguments{
\item{data}{\code{n} by \code{(p+2)} data frame in which the first column is subject IDs, the second column is
the time points of longitudinal data.}

\item{lambda}{numerical vector of tuning parameters. For one-stage model, \code{lambda} is a scalar controlling the
sparsity of the network. For two-stage model,
\code{lambda} is a vector of length 2, in which the first entry controls the sparsity of both
pre-treatment and post-treatment networks,
while the second entry controls the overlap of the two networks.}

\item{group}{factor  of length \code{n} if supplied. If \code{group} is a  one-level factor,
then a one-stage model is fitted. If \code{group} is a two-level factor, then a two-stage model is fitted.
Data points that are before (after) the treatment (exposure) share the same
level in \code{group}. If NULL, then a one-stage model is fitted.}

\item{random}{a logical variable. If TRUE, then a heterogeneous model is fitted.
Otherwise, a homogeneous model is fitted.}

\item{expFix}{numerical number specifying the exponent in the covariance function of
the longitudinal data. Default is 1.}

\item{N}{a integer specifying the number of sampling for heterogeneous model}

\item{maxit}{integer specifying  the maximum iterations for the algorithms.}

\item{tol}{a small number determining whether the algorithms converged and should stop.}

\item{lower}{vector of length 1 or 2 which specifies the lower bounds for temporal correlation \code{tau}.
It is of length 1 for one-stage model and 2 for two-stage model}

\item{upper}{vector of length 1 or 2 which specifies the upper bounds for temporal correlation \code{tau}
It is of length 1 for one-stage model and 2 for two-stage model}

\item{w.init}{initial value for covariance matrix. Default is identity matrix.}

\item{wi.init}{initial value for precision matrix. Default is identity matrix.}

\item{trace}{whether or not show the progress of the computation}

\item{...}{other inputs}
}
\value{
A list which includes:

\code{w:} the list of the estimates for covariance matrices

\code{wi:} the list of the estimates for precision matrices

\code{tau:} the estimate of dampening rate \code{tau}. For heterogeneous models, the
output is a vector called random effects. For homogeneous model, the output is a scalar.

\code{alpha:} a scalar representing the estimate of the parameter in exponential distribution of \code{tau} for heterogeneous models.
The output is NULL for homogeneous model

\code{ll:} the likelihood.  \code{ll} is used to compute extended QIC for tuning parameter selection
}
\description{
This function estimates  precision matrices(networks) and random effects from longitudinal
high-dimensional data under normality assumption.
}
\details{
\code{lglasso} is developed to estimate precision matrices, or networks, and random effects \eqn{\tau_i}
from high-dimensional longitudinal data.
The one-stage model in \code{lglasso} is proposed in Zhou \emph{et al} (2024).
Currently, it contains two network identification models,
\emph{i.e.,} one-stage  and two-stage model.
One-stage model assume a common network underlying
the longitudinal data for all the subjects. Consequently,
the function outputs a single network as the estimate in this case.
In two-stage models, a treatment is applied at time \eqn{t_i} during the time interval
for subject \eqn{i}.
Therefore, there are two networks, \emph{i.e}., pre- and post-treatment networks,  that need to be estimated simultaneously.
For details, please check the reference paper
and the online resources.
}
\examples{
\dontrun{
library(lglasso)
# number of nodes
p=15
# number of edge in general network
m1=20
# the difference between the number of edges in individual networks and general network
m2=5
# number of subjects
n=30
set.seed(1)
## One-stage model
### Estimate the network based on homogeneous one-stage model
####simulate data
dd=Simulate(type="homo",n=n,p=p,m1=m1,m2=m2,tau=2,tt=10)
ddata=dd$data
dim(ddata)
ddata[1:2,1:5]
#### Estimation
aa=lglasso(data=ddata,lambda = 0.01,trace=TRUE)
estimates=lapply(aa$wi,function(ll){ifelse(abs(ll)>10^(-5),1,0)})
#### estimated network
estimates
#### true network
dd$network
#### correlation parameter
aa$tau
#### likelihood
aa$ll



### Estimate the network based on heterogeneous one-stage model
####simulate data
dd=Simulate(type="heter",n=n,p=p,m1=m1,m2=m2,tt=10,alpha=5,group = 1)
ddata=dd$data$pre
dim(ddata)
ddata[1:2,1:5]
#### Estimation
aa=lglasso(data=ddata,lambda = 0.01,random=TRUE,trace=TRUE)
estimates=lapply(aa$wi,function(ll){ifelse(abs(ll)>10^(-5),1,0)})
#### estimated network
estimates
####  true network
dd$network
#### likelihood
aa$ll



## Two-stage model
### Estimate the networks based on homogeneous two-stage model
####simulate data
dd=Simulate(type="homo",n=n,p=p,m1=m1,m2=m2,tau=c(2,1),tt=10)
ddata=do.call(rbind,dd$data)
group=c(rep(0,nrow(ddata)/2),rep(1,nrow(ddata)/2))
dim(ddata)
ddata[1:2,1:5]
#### Estimation
aa=lglasso(data=ddata,lambda = c(0.01,0.01),group = group,trace=TRUE)
estimates=lapply(aa$wi,function(ll){ifelse(abs(ll)>10^(-5),1,0)})
#### estimated pre-treatment  network
estimates[[1]]
####  true pre-treatment network
dd$network$pre
#### estimated post-treatment  network
estimates[[2]]
####  true post-treatment network
dd$network$post
#### correlation parameters
aa$tau
#### likelihood
aa$ll


### Estimate the networks based on heterogeneous two-stage model
####simulate data
dd=Simulate(type="heter",n=n,p=p,m1=m1,m2=m2,tt=10,alpha=0.5,group=2)
ddata=do.call(rbind,dd$data)
group=c(rep(0,nrow(ddata)/2),rep(1,nrow(ddata)/2))
dim(ddata)
ddata[1:2,1:5]
#### Estimation
aa=lglasso(data=ddata,lambda = c(0.01,0.01),random=TRUE,group = group,trace=TRUE)
estimates=lapply(aa$wi,function(ll){ifelse(abs(ll)>10^(-5),1,0)})
#### estimated pre-treatment network
estimates[[1]]
#### true pre-treatment network
dd$network$pre
#### estimated post-treatment network
estimates[[2]]
####  true post-treatment network
dd$network$post
#### likelihood
aa$ll
}
}
