% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_functions.R
\name{lli_homo}
\alias{lli_homo}
\title{full log likelihood used in EBIC computation}
\usage{
lli_homo(idata, omega, tau, type)
}
\arguments{
\item{idata}{Data matrix for the subject i in which the first column is id for subject, the second column is
the time points of observation.  Columns 2 to (p+2) is the observations for p variables.}

\item{omega}{Precision matrix}

\item{tau}{Correlation parameter}

\item{type}{Type of correlation function, which can take either  "abs" or "qua".}
}
\value{
Value of likelihood function for subject i at given omega and tau
}
\description{
full log likelihood used in EBIC computation
}
\author{
Jie Zhou
}
