% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heter.R
\name{logdensity}
\alias{logdensity}
\title{Negative likelihood function used in EM algorithm of heterogeneous marginal graphical lasso model}
\usage{
logdensity(idata, omega, tau, alpha, type)
}
\arguments{
\item{idata}{Data matrix for the subject i in which the first column is id for subject, the second column is
the time points of observation.  Columns 2 to (p+2) is the observations for p variables.}

\item{omega}{Precision matrix}

\item{tau}{Correlation parameter}

\item{alpha}{Parameter in exponential distribution}

\item{type}{Type of correlation function, which can take either  "abs" or "qua".}
}
\value{
Value of complete likelihood function at given value of omega, tau and alpha
}
\description{
Negative likelihood function used in EM algorithm of heterogeneous marginal graphical lasso model
}
\author{
Jie Zhou
}
