% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_functions.R
\name{mle_alpha}
\alias{mle_alpha}
\title{Maximum likelihood estimate of correlation parameter for given structure of precision matrix}
\usage{
mle_alpha(data, alpha0, omega, type, tole, lower, upper)
}
\arguments{
\item{data}{Data matrix  in which the first column is subject id, the second column is
the time points of observation.  Columns 2 to (p+2) is the observations for p variables.}

\item{alpha0}{Initial value for the parameter in exponential distribution}

\item{omega}{Fixed value for precision matrix}

\item{type}{Type of correlation function, which can take either  "abs" or "qua".}

\item{tole}{Error tolerance for determination of convergence of EM algorithm}

\item{lower}{Lower bound for prediction of correlation parameter tau}

\item{upper}{Upper bound for prediction of correlation parameter tau}
}
\description{
Maximum likelihood estimate of correlation parameter for given structure of precision matrix
}
\author{
Jie Zhou
}
