% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{mle_net}
\alias{mle_net}
\title{Title}
\usage{
mle_net(data, priori)
}
\arguments{
\item{data}{A Longitudinal data set}

\item{priori}{Given structure of precision matrix}
}
\value{
The maximum likelihood estimation
}
\description{
Title
}
\author{
Jie Zhou
}
