% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_compare.R
\name{power_compare}
\alias{power_compare}
\title{Main simulation function comparing five methods for network seletion.}
\usage{
power_compare(m, n, p, coe, l, rho, prob, heter)
}
\arguments{
\item{m}{number of subjects}

\item{n}{number of observations per subjects}

\item{p}{the dimension of the data to be generated}

\item{coe}{coefficients for covariates}

\item{l}{the simulation scale}

\item{rho}{tuning parameter for glasso}
}
\value{
list with length equal to 5.
}
\description{
Main simulation function comparing five methods for network seletion.
}
