% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_compare1.R
\name{power_compare1}
\alias{power_compare1}
\title{Main simulation function comparing five methods for network selection.}
\usage{
power_compare1(
  m,
  n,
  p,
  coe,
  l,
  rho,
  prob,
  heter,
  community2 = F,
  uu = c(0, 0),
  zirate = c(0.2, 0)
)
}
\arguments{
\item{m}{number of subjects}

\item{n}{number of observations per subjects}

\item{p}{the dimension of the data to be generated}

\item{coe}{coefficients for covariates}

\item{l}{the simulation scale}

\item{uu}{uu[1] is the dampening rate for community 1,
uu[2] is the dampening rate for community 2.Both are in log scale and
in alpha form.}

\item{rho1}{tuning parameter for glasso}

\item{rho2}{tuning parameter for glasso}
}
\value{
list with length equal to 3.
}
\description{
Main simulation function comparing five methods for network selection.
}
