% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_compare2.R
\name{power_compare2}
\alias{power_compare2}
\title{Title}
\usage{
power_compare2(m, n, p, coe, l, rho, prob, heter)
}
\arguments{
\item{m}{number of subjects}

\item{n}{number of observations per subjects}

\item{p}{the dimension of the data to be generated}

\item{coe}{coefficients for covariates}

\item{l}{the simulation scale}

\item{rho}{tuning parameter for glasso}

\item{prob}{the density of the network}

\item{heter}{indicator}
}
\value{
list
}
\description{
Title
}
