% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{sim_2heter}
\alias{sim_2heter}
\title{Generate the heterogeneous data}
\usage{
sim_2heter(p, prob, alpha1, alpha2, age)
}
\arguments{
\item{p}{An integer indicating dimension of the network}

\item{prob}{The density of the edges in the network}

\item{age}{A list with length equal to the number of subjects. Each component is a vector indicating the
time points at which the observations was made.}

\item{alpha}{A positive number representing the parameter for the exponential distribution}
}
\value{
A list for the simulated data and its parameters
}
\description{
Generate the heterogeneous data
}
