% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{sim_2homo}
\alias{sim_2homo}
\title{Generate the two-community homogeneous data}
\usage{
sim_2homo(p, prob, tau1, tau2, age, zirate = c(0.2, 0))
}
\arguments{
\item{p}{An integer indicating dimension of the network}

\item{prob}{The density of the edges in the network}

\item{tau1}{The shared correlation among observations in first community.}

\item{tau2}{The correlation in second community}

\item{age}{A list with length equal to the number of subjects. Each component is a vector indicating the
time points at which the observations was made.}
}
\value{
A list for simulated data
}
\description{
Generate the two-community homogeneous data
}
