% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{sim_homo}
\alias{sim_homo}
\title{Generate the homogeneous data}
\usage{
sim_homo(p, prob, tau, age)
}
\arguments{
\item{p}{An integer indicating dimension of the network}

\item{prob}{The density of the edges in the network}

\item{tau}{The shared correlation among observations.}

\item{age}{A list with length equal to the number of subjects. Each component is a vector indicating the
time points at which the observations was made.}
}
\value{
A list for simulated data
}
\description{
Generate the homogeneous data
}
