% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{AA}
\alias{AA}
\title{Title}
\usage{
AA(
  B,
  data,
  type = c("expFixed"),
  expFix,
  maxit = 30,
  tol = 10^(-4),
  lower = c(0.01, 0.1),
  upper = c(10, 5),
  ...
)
}
\arguments{
\item{B}{a p by p given precision matrix}

\item{data}{a (p+2)-by-n data frame}

\item{type}{specify how to model the covariance matrix}

\item{expFix}{the parameter in variance function when the data are longitudinal}

\item{maxit}{the maximum iteration number}

\item{tol}{the minimum difference for the algorithm to be called converged}

\item{lower}{vector of length 1 or 2 which specifies the lower bounds for alpha_1 (and alpha_2) in the correlation matrix}

\item{upper}{vector of length 1 or 2 which specifies the upper bounds for alpha_1 (and alpha_2) in the correlation matrix}

\item{...}{other unspecifed parameterss}
}
\value{
a list of matrices
}
\description{
Title
}
