% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{AAheter}
\alias{AAheter}
\title{Estimate the phimatrix in heterogeneous model}
\usage{
AAheter(data, wi, alpha, group, l = 5000, expFix = 1, ...)
}
\arguments{
\item{data}{longitudinal data set}

\item{wi}{given precision matrix}

\item{alpha}{exponential distribution with rate alpha}

\item{group}{specify how data is grouped}

\item{l}{number of random samples in importance sampling}
}
\value{
a list for estimates of tau and AA
}
\description{
Estimate the phimatrix in heterogeneous model
}
