% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{BB}
\alias{BB}
\title{Title}
\usage{
BB(A, data, lambda, random = FALSE, tau)
}
\arguments{
\item{A}{a list of length 1 or 2 corresponding to the number of stages. The each entry of A is a
list representing all phi matrices before or after the treatment.}

\item{data}{a list of (p+2)-by-ni data frame}

\item{lambda}{given tuning parameter(s)}

\item{random}{logical variable indicating the type of the model}

\item{tau}{scalar if \emph{random} is FALSE and a vector if \emph{random} is TRUE}
}
\value{
a list with the same length as A
}
\description{
Title
}
