% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{BB}
\alias{BB}
\title{Title}
\usage{
BB(A, data, lambda, type = c("expFixed"), diagonal = TRUE)
}
\arguments{
\item{A}{a list of length 1 or 2 corresponding to the number of phases. The each entry of A is a
list representing all phi matrices before or after the treatment.}

\item{data}{a list of (p+2)-by-ni data frame}

\item{lambda}{given tuning parameter(s)}

\item{type}{specify model type}
}
\value{
a list with the same length as A
}
\description{
Title
}
