% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{BB}
\alias{BB}
\title{Title}
\usage{
BB(A, data, lambda, type = c("expFixed"), diagonal = TRUE)
}
\arguments{
\item{A}{a list representing the phimatrices}

\item{data}{a (p+2)-by-n data frame}

\item{lambda}{given tuning parameter(s)}

\item{type}{specify model type}
}
\value{
a list with the same length as A
}
\description{
Title
}
