% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{BB}
\alias{BB}
\title{Title}
\usage{
BB(
  A,
  data,
  lambda,
  type = c("expFixed"),
  diagonal = TRUE,
  maxit = 100,
  tol = 10^(-4),
  lower = c(0.01, 0.1),
  upper = c(10, 5),
  start = c("warm", "cold"),
  w.init = NULL,
  wi.init = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
