% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{comparison}
\alias{comparison}
\title{Comparison function}
\usage{
comparison(real, estimate)
}
\arguments{
\item{real}{The real network}

\item{estimate}{The estimated network}
}
\value{
TPR and FPR
}
\description{
Comparison function
}
