% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{cvError}
\alias{cvError}
\title{Title}
\usage{
cvError(data.train, data.valid, BB, group.train = NULL, group.valid = NULL)
}
\arguments{
\item{data.train}{trainng data}

\item{data.valid}{testing data}

\item{BB}{given network}

\item{group.train}{group in training data}

\item{group.valid}{group in testing data}
}
\value{
a matrix
}
\description{
Title
}
