% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{cvError}
\alias{cvError}
\title{Compute the cross validation error}
\usage{
cvError(data.train, data.valid, B, group.train = NULL, group.valid = NULL)
}
\arguments{
\item{data.train}{trainng data}

\item{data.valid}{testing data}

\item{B}{given network (or network list)}

\item{group.train}{group in training data}

\item{group.valid}{group in testing data}
}
\value{
a matrix
}
\description{
Compute the cross validation error
}
