% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{CVlglasso}
\alias{CVlglasso}
\title{Cross validation for \code{lglasso}}
\usage{
CVlglasso(
  data,
  K,
  group = NULL,
  random = FALSE,
  lambda = NULL,
  nlam = 10,
  lam.min.ratio = 0.01,
  expFix = 1,
  trace = FALSE
)
}
\arguments{
\item{data}{same as in \code{lglasso}}

\item{K}{fold number for the cross validation}

\item{group}{same as in \code{lglasso}}

\item{random}{same as in \code{lglasso}}

\item{lambda}{tuning parameter. For one-stage model, lambda is a vector. For two-stage model,
lambda is a 2-column matrix. The first column  is tuning parameter controlling the sparsity,
the second column is tuning parameter controlling the
similarity of two networks.}

\item{nlam}{If \code{lambda} is NULL, then \code{nlam} set the number of tuning parameter}

\item{lam.min.ratio}{ratio of largest lambda vs smallest lambda}

\item{expFix}{same as in \code{lglasso}}

\item{trace}{logical variable. Whether show the computation process}
}
\value{
list of length 2.  The first component is the cross validation errors, the second component is the corresponding
tuning parameters
}
\description{
The function computes the cross validation errors  in \code{lglasso}.
}
