% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{CVlglasso}
\alias{CVlglasso}
\title{Cross validation for \code{lglasso}}
\usage{
CVlglasso(
  data,
  group = NULL,
  random = FALSE,
  lambda = NULL,
  nlam = 10,
  lam.min.ratio = 0.01,
  K,
  expFix = 1,
  trace = FALSE,
  NN = 500
)
}
\arguments{
\item{data}{raw data}

\item{group}{group variable}

\item{random}{a logical variable indicating the type of the model}

\item{lambda}{tuning parameter}

\item{nlam}{number of tuning parameter}

\item{lam.min.ratio}{ratio of largest lambda vs smallest lambda}

\item{K}{cv folds}

\item{expFix}{given parameter}

\item{trace}{whether show the process}

\item{NN}{the number of sampling}
}
\value{
list of which the first component is the cross validation errors and the second component is the corresponding
tuning parameters
}
\description{
The function computes the cross validation errors for one of the three network models in \code{lglasso} command.
}
