% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{CVlglasso}
\alias{CVlglasso}
\title{Cross validation for \code{lglasso}}
\usage{
CVlglasso(
  type = c("expFixed"),
  data,
  group = NULL,
  random = FALSE,
  lambda = NULL,
  nlam = 10,
  lam.min.ratio = 0.01,
  K,
  expFix = 1,
  trace = FALSE,
  NN = 500
)
}
\arguments{
\item{type}{underlying model type, either \code{general}, \code{longihomo} or \code{longiheter}.}

\item{data}{raw data}

\item{group}{group variable}

\item{lambda}{tuning parameter}

\item{nlam}{number of tuning parameter}

\item{lam.min.ratio}{ratio of largest lambda vs smallest lambda}

\item{K}{cv folds}

\item{expFix}{given parameter}

\item{trace}{whether show the process}

\item{cores}{parallel computing}
}
\value{
list of which the first component is the cross validation errors and the second component is the corresponding
tuning parameters
}
\description{
The function computes the cross validation errors for one of the three network models in \code{lglasso} command.
}
\section{crit.cv = match.arg(crit.cv)}{

}

\section{start = match.arg(start)}{
Sminus = S
diag(Sminus) = 0
if (is.null(lambda)) {
if (!((lam.min.ratio <= 1) && (lam.min.ratio > 0))) {
cat("\nlam.min.ratio must be in (0, 1]... setting to 1e-2!")
lam.min.ratio = 0.01
}
if (!((nlam > 0) && (nlam\%\%1 == 0))) {
cat("\nnlam must be a positive integer... setting to 10!")
nlam = 10
}
lam.max = max(abs(Sminus))
lam.min = lam.min.ratio * lam.max
lambda = 10^seq(log10(lam.min), log10(lam.max), length = nlam)
if (!is.null(group)){
lambda=expand.grid(lambda,lambda)
}
}
else {
if (is.null(group)){
lambda = sort(lambda)
}
}

nnlambda=ifelse(is.null(group),length(lambda),nrow(lambda))
cv_error=matrix(0,nrow=nnlambda,ncol=K)

k=NULL
k=1
CV = foreach(k = 1:K, .packages = "lglasso", .combine = "cbind",
.inorder = FALSE) \%dopar\% {

\if{html}{\out{<div class="sourceCode">}}\preformatted{             if (trace) \{
               progress = utils::txtProgressBar(max = K, style = 3)
             \}

             leave.out =subjects[ind[(1 + floor((k - 1) * n/K)):floor(k *
                                                                        n/K)]]
             indexValid=which(data[,1] \%in\% leave.out)
             data.train = data[-indexValid, , drop = FALSE]
             data_bar = apply(data.train[,-c(1,2)], 2, mean)
             data.train[,-c(1,2)] = scale(data.train[,-c(1,2)], center = data_bar, scale = FALSE)
             data.valid = data[indexValid,, drop = FALSE]
             data.valid[,-c(1,2)] = scale(data.valid[,-c(1,2)], center = data_bar, scale = FALSE)
             group.train=group[-indexValid]
             group.valid=group[indexValid]
             #S.train = crossprod(data.train[,-c(1,2)])/(dim(data.train)[1])
             #S.valid = crossprod(data.valid[,-c(1,2)])/(dim(data.valid)[1])


             if (is.null(group))\{

               aa= sapply(lambda, function(x) \{lglasso(data=data.train,lambda=x)$wi\})
               cc=lapply(aa, function(B)\{
                 M=ifelse(abs(B)<=10^(-2), 0,1)
                 diag(M)=2
                 M
               \})
               bb=lapply(cc, function(B) \{cvError(data.train=data.train,data.valid=data.valid,B=B)\})
               bb=do.call(c,bb)
             \}
             if (!is.null(group))\{
               aa= apply(lambda,1, function(x) \{lglasso(data=data.train,lambda=x, expFix = expFix, group=group.train)$wi\})
               cc=lapply(aa, function(B)\{
                 lapply(B, function(Z)\{
                   M=ifelse(abs(Z)<=10^(-1), 0,1)
                   diag(M)=2
                   M
                 \}
                 )
               \}
               )

               bb=lapply(cc, function(BB)\{cvError(data.train=data.train,data.valid=data.valid,B=BB,
                                                  group.valid=group.valid, group.train = group.train)\})
               bb=do.call(c,bb)
             \}

             cv_error=bb

             if (trace) \{
               utils::setTxtProgressBar(progress,  k)
             \}

             return(cv_error=cv_error)
           \}
}\if{html}{\out{</div>}}

stopCluster(cluster)
output=structure(list(cv_error=CV, lambda=lambda),class="cvlglasso")
return(output)
}
}

