% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{cvlglassofull}
\alias{cvlglassofull}
\title{Title}
\usage{
cvlglassofull(
  data,
  group = NULL,
  lambda = NULL,
  random = FALSE,
  nlam = 10,
  lam.min.ratio = 0.01,
  K,
  expFix = 1,
  trace = FALSE
)
}
\arguments{
\item{data}{raw data}

\item{group}{group variable}

\item{lambda}{tuning parameter}

\item{nlam}{number of tuning parameter}

\item{lam.min.ratio}{ratio of largest lambda vs smallest lambda}

\item{K}{cv folds}

\item{expFix}{given parameter}

\item{trace}{whether show the process}

\item{type}{model type}

\item{cores}{parallel computing}
}
\value{
list
}
\description{
Title
}
