% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{importanceEstimates}
\alias{importanceEstimates}
\title{Title}
\usage{
importanceEstimates(importancesSample, datai, groupi)
}
\arguments{
\item{importancesSample}{the random samples}

\item{datai}{the data for subject i}

\item{groupi}{specify how datai is grouped}
}
\value{
a list of estimated
}
\description{
Title
}
