% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{iss}
\alias{iss}
\title{Quasi covariance matrix for subject i}
\usage{
iss(idata, itau, type)
}
\arguments{
\item{idata}{Data matrix for the subject i in which the first column is subject (cluster) id, the second column stands for
the time points () of observation.  Columns 2 to (p+2) is the observations for p variables respectively.}

\item{itau}{Correlation parameter}

\item{type}{Type of correlation function, which typically take either  0, 1 or 2.}
}
\value{
Empirical quasi covariance matrix
}
\description{
Quasi covariance matrix for subject i
}
\author{
Jie Zhou
}
