% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lglasso-package.R
\docType{package}
\name{lglasso-package}
\alias{lglasso-package}
\title{lglasso: Graphical Lasso for Longitudinal Data}
\description{
Estimate treatment-specific precision matrices (networks) from longitudinal high-dimensional normal data. The corresponding random effects are also estimated. It is motivated by the analysis of omics data in clinical trials where the longitudinal omics data becomes increasingly common. It includes both one-stage models (without treatment) and two-stage models (with one treatment). For details of the algorithms, please check the materials on its GitHub repo. If you have any questions, feel free to contact the maintainers through the email below.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jiezhou-2/lglasso}
}

}
\author{
\strong{Maintainer}: Jie Zhou \email{chowstat@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Jiang Gui
  \item Weston Viles
  \item Anne Hoen
}

}
\keyword{internal}
