% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{lglasso}
\alias{lglasso}
\title{Longitudinal graphical lasso}
\usage{
lglasso(
  data,
  lambda,
  group = NULL,
  random = FALSE,
  expFix = 1,
  N = 100,
  maxit = 50,
  tol = 10^(-2),
  lower = c(0.01, 0.01),
  upper = c(10, 10),
  w.init = NULL,
  wi.init = NULL,
  trace = FALSE,
  ...
)
}
\arguments{
\item{data}{\code{n} by \code{(p+2)} data frame in which the first column is for subject IDs, the second column is
for the time points of longitudinal data.}

\item{lambda}{numerical vector of tuning parameters,controlling the
sparsity of the networks. For one-stage model, \code{lambda} is a scalar. For two-stage model,
\code{lambda} is a vector of length 2.  For details, see the explanations in the below.}

\item{group}{factor  of length \code{n} if supplied. It indicates each data
point either before or after the treatment (exposure). Default is a single-level factor,
which corresponds to the one-stage model. For two-stage model, \code{group} is a two-level factor.
where the two levels correspond to the two stage.}

\item{random}{a logical variable. If TRUE, then a heterogeneous model is estimated.
Otherwise, a homogeneous model is estimated.}

\item{expFix}{numerical number to specify the form of the covariance function of the longitudinal data.}

\item{N}{a integer specifying the number of sampling for heterogeneous model}

\item{maxit}{integer of  the maximum iterations for the algorithms.}

\item{tol}{value for determining if the algorithms have converged.}

\item{lower}{vector of length 1 or 2 which specifies the lower bounds for temporal correlation parameter alpha_1 (and alpha_2) in the correlation matrix}

\item{upper}{vector of length 1 or 2 which specifies the upper bounds for temporal correlation parameter alpha_1 (and alpha_2) in the correlation matrix}

\item{w.init}{initial value for covariance matrix}

\item{wi.init}{inital value for precision matrix}

\item{trace}{whether or not show the progress of the computation}

\item{...}{other inputs}
}
\value{
list which include following components:

\code{w} the list of the estimates for covariance matrices

\code{wi} the list of the estimates for precision matrices

\code{tau} the estimate of dampening rate \code{tau}

\code{alpha} the parameter in exponential distribution of \code{tau} for heterogeneous models

\code{ll} the value of likelihood.  \code{ll} can be used to compute Extended BIC for tuning parameter selection
}
\description{
This is the main function of the package which estimates the underlying precision matrices(networks) from longitudinal
high-dimensional data under normality assumption.
}
\details{
\code{lglasso} is the main function of the package which aims to estimate precision matrices, or networks, from longitudinal data.
It is based on the models in Zhou \emph{et al} (2024).
Currently, it contains two network identification models,
\emph{i.e.,} one-stage  and two-stage model.
One-stage model assume a common network underlying
the longitudinal data for all the subjects. Consequently,
\code{lglasso} only outputs a single network as the estimate.
Two-stage model allows that a treatment occurred at time point \eqn{t_i}
for subject \eqn{i}.
Therefore, there are two networks, i.e., pre- and post-treatment networks,  that need to be estimated.
The core idea behind the models in \code{lglasso} is that the models decompose the covariance matrix
of longitudinal data into
temporal and cross-section part, where the model for the temporal part has the form of
\deqn{\exp(-|t_{i}-t_{j}|^{(-\tau)}}. For details, please check the  paper in the reference.
}
