% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{lglasso}
\alias{lglasso}
\title{Longitudinal graphical lasso}
\usage{
lglasso(
  data,
  lambda,
  group = NULL,
  random = FALSE,
  expFix = 1,
  N = 100,
  maxit = 50,
  tol = 10^(-2),
  lower = c(0.01, 0.1),
  upper = c(10, 5),
  w.init = NULL,
  wi.init = NULL,
  trace = FALSE,
  ...
)
}
\arguments{
\item{data}{a \code{n} by \code{(p+2)} data frame in which the first column is subject ID, the second column is
the time point for longitudinal data or tissue ID.}

\item{lambda}{vector of length 1 or 2,  which
is the tuning parameter for the identification of the networks. For details, see the explanations in the below.}

\item{group}{vector  of length \code{n} if supplied which specify which data
points need to be grouped together to infer the heterogeneous networks for, e.g, pre/post vaccination.}

\item{random}{a logical variable specifying the type of the model}

\item{expFix}{numeric number used in the model specification}

\item{N}{a integer specifying the number of sampling for heterogeneous model}

\item{maxit}{the maximum iterations for the estimation.}

\item{tol}{the minimum value for  convergence criterion}

\item{lower}{vector of length 1 or 2 which specifies the lower bounds for alpha_1 (and alpha_2) in the correlation matrix}

\item{upper}{vector of length 1 or 2 which specifies the upper bounds for alpha_1 (and alpha_2) in the correlation matrix}

\item{w.init}{initial value for covariance matrix}

\item{wi.init}{inital value for precision matrix}

\item{trace}{whether or not show the progress of the computation}

\item{...}{other inputs}
}
\value{
list which include following components:

\code{w} the general covariance matrix estimate;

\code{wList} list representing the individual covariance matrix estimate;

\code{wi} the general precision matrix estimate;

\code{wiList} list representing the individual precision matrix estimate;

\code{v} the correlation matrix between specified classes;

\code{vList} list representing the individual correlation matrix;

\code{tauhat} the correlation parameters for longitudinal data
}
\description{
This is the main function of the package, which identifies the underlying  network model from clustered data. Here
clustered data include longitudinal data, or spatially correlated data,
e.g, metabolites in different tissues of a same subject.  .
or more broadly, clustered data for given tuning parameters.
}
\details{
This function is the main function of the package. It is
designed to identify networks from longitudinal data.
It implements two network identification models, \emph{i.e.,} one-stage (two-stage) model.
One-stage model assume a common network underlying
the longitudinal data for all the subjects. Consequently,
\emph{lglasso} only outputs a single network as the estimate.
Two-stage model assume a change time point \emph{t_i} that occurs for subject \emph{i(1<= i <= m)}.
The networks before and after \emph{t_i} are different.  Let's say we have
two time points,t_i,t_j, then in model \code{general}, the correlation is
tau_ij, while in model \emph{expFixed}, we have  tau=exp(-alpha_1|t_1-t_2|^(-alpha_2))
with alpha_2 need to be pre-specified (default is alpha_2=1). In model \code{twoPara},
both alpha_1 and alpha_2 is unknown and need to be inferred from the data.
For longitudinal data, model \code{expFixed} is recommended while for omics data
from different tissues or contents, model \code{general} should be adopted.
}
