% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{lglasso}
\alias{lglasso}
\title{Longitudinal graphical lasso}
\usage{
lglasso(
  data,
  lambda,
  type = c("general", "expFixed", "twoPara"),
  expFix = 1,
  group = NULL,
  diagonal = TRUE,
  maxit = 30,
  tol = 10^(-3),
  lower = c(0.01, 0.1),
  upper = c(10, 5),
  start = c("cold", "warm"),
  w.init = NULL,
  wi.init = NULL,
  trace = FALSE,
  ...
)
}
\arguments{
\item{data}{a \code{n} by \code{(p+2)} data frame in which the first column is subject ID, the second column is
the time point for longitudinal data or tissue ID.}

\item{lambda}{vector of length 1 or 2,  which
is the tuning parameter for the identification of the networks. For details, see the explanations in the below.}

\item{type}{a string specifying which model need to be fitted. There are three models available,
which are refered as \code{general, expFixed} and \code{twoPara} respectively.
 Please see the details in the below for the meaning of each.}

\item{expFix}{numeric number used in the model specification}

\item{group}{vector  of length \code{n} if supplied which specify which data
points need to be grouped together to infer a network for each of them.}

\item{maxit}{the maximum iterations for the estimation.}

\item{tol}{the minimum value for  convergence criterion}

\item{lower}{vector of length 1 or 2 which specifies the lower bounds for the optimization algorithm}

\item{upper}{vector of length 1 or 2 which specifies the upper bounds for the optimization algorithm}

\item{...}{other inputs}
}
\value{
\code{w} the general covariance matrix estimate

\code{wList} list representing the individual covariance matrix estimate

\code{wi} the general precision matrix estimate

\code{wiList} list representing the individual precision matrix estimate

\code{v} the correlation matrix between specified classes

\code{vList} list representing the individual correlation matrix

\code{tauhat} the correlation parameters for longitudinal data
}
\description{
This is the main function of the package, which identifies the underlying  network model from clustered data. Here
 clustered data include longitudinal data, or spatially correlated data,
 e.g, metabolites in different tissues of a same subject.  .
 or more broadly, clustered data for given tuning parameters.
}
