% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{lglasso}
\alias{lglasso}
\title{Longitudinal graphical lasso}
\usage{
lglasso(
  data,
  lambda,
  group = NULL,
  random = FALSE,
  expFix = 1,
  N = 100,
  maxit = 30,
  tol = 10^(-1),
  lower = c(0.01, 0.1),
  upper = c(10, 5),
  start = c("cold", "warm"),
  w.init = NULL,
  wi.init = NULL,
  trace = FALSE,
  type = c("expFixed"),
  ...
)
}
\arguments{
\item{data}{a \code{n} by \code{(p+2)} data frame in which the first column is subject ID, the second column is
the time point for longitudinal data or tissue ID.}

\item{lambda}{vector of length 1 or 2,  which
is the tuning parameter for the identification of the networks. For details, see the explanations in the below.}

\item{group}{vector  of length \code{n} if supplied which specify which data
points need to be grouped together to infer the heterogeneous networks for, e.g, pre/post vaccination.}

\item{expFix}{numeric number used in the model specification}

\item{maxit}{the maximum iterations for the estimation.}

\item{tol}{the minimum value for  convergence criterion}

\item{lower}{vector of length 1 or 2 which specifies the lower bounds for alpha_1 (and alpha_2) in the correlation matrix}

\item{upper}{vector of length 1 or 2 which specifies the upper bounds for alpha_1 (and alpha_2) in the correlation matrix}

\item{start}{how to start the initial values for lglasso algorithm}

\item{w.init}{initial value for covariance matrix}

\item{wi.init}{inital value for precision matrix}

\item{trace}{whether or not show the progress of the computation}

\item{type}{a string specifying which model need to be fitted. There are three models available,
which are referred as \code{general}, \code{expFixed} respectively.
Please see the details in the below for the meaning of each.}

\item{...}{other inputs}
}
\value{
list which include following components:

\code{w} the general covariance matrix estimate;

\code{wList} list representing the individual covariance matrix estimate;

\code{wi} the general precision matrix estimate;

\code{wiList} list representing the individual precision matrix estimate;

\code{v} the correlation matrix between specified classes;

\code{vList} list representing the individual correlation matrix;

\code{tauhat} the correlation parameters for longitudinal data
}
\description{
This is the main function of the package, which identifies the underlying  network model from clustered data. Here
clustered data include longitudinal data, or spatially correlated data,
e.g, metabolites in different tissues of a same subject.  .
or more broadly, clustered data for given tuning parameters.
}
\details{
This function implements three statistical models for  network inference,
according to how the correlations is specified between time points (or tissues or
contents in some clinical studies). These three models are referred as
\code{general}, \code{expFixed}.Let's say we have
two time points,t_i,t_j, then in model \code{general}, the correlation is
tau_ij, while in model \emph{expFixed}, we have  tau=exp(-alpha_1|t_1-t_2|^(-alpha_2))
with alpha_2 need to be pre-specified (default is alpha_2=1). In model \code{twoPara},
both alpha_1 and alpha_2 is unknown and need to be inferred from the data.
For longitudinal data, model \code{expFixed} is recommended while for omics data
from different tissues or contents, model \code{general} should be adopted.
}
