% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{lglassoHeter}
\alias{lglassoHeter}
\title{Title}
\usage{
lglassoHeter(
  data,
  lambda,
  group,
  maxit,
  tol = 10^(-3),
  trace = FALSE,
  start = c("warm", "cold"),
  w.init = NULL,
  wi.init = NULL,
  N,
  expFix = 1,
  ...
)
}
\arguments{
\item{trace}{}
}
\description{
Title
}
