% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_functions.R
\name{ll_homo}
\alias{ll_homo}
\title{Value of likelihood function at given parameter}
\usage{
ll_homo(data, omega, tau, type)
}
\arguments{
\item{data}{Data matrix  in which the first column is subject id, the second column is
the time points of observation.  Columns 2 to (p+2) is the observations for p variables.}

\item{omega}{Precision matrix}

\item{tau}{Correlation parameter}

\item{type}{Type of correlation function, which can take either  "abs" or "qua".}
}
\value{
Value of likelihood function at given omega and tau
}
\description{
Value of likelihood function at given parameter
}
\author{
Jie Zhou
}
