% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{phifunction}
\alias{phifunction}
\title{Title}
\usage{
phifunction(t, tau, expFix = 1)
}
\arguments{
\item{t}{a vector specify the time points corresponding to the data}

\item{tau}{the damping rate parameter with length 1 or 2}

\item{expFix}{a scalar specifying the form of correlation function}
}
\value{
a square matrix used to construct the likelihood
}
\description{
Title
}
