% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_functions.R
\name{phifunction}
\alias{phifunction}
\title{Construct the temporal component fo correlation function}
\usage{
phifunction(t, tau, type = 1)
}
\arguments{
\item{t}{Time points of observations}

\item{tau}{correlation parameter}

\item{type}{The type of correlation function, which typically take either 0,1 or 2.}
}
\value{
A square matrix with dimension equal to the length of vector t
}
\description{
Construct the temporal component fo correlation function
}
\author{
Jie Zhou
}
