% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{plot.cvlglasso}
\alias{plot.cvlglasso}
\title{Title}
\usage{
\method{plot}{cvlglasso}(x, xvar = c("lambda", "step"), ...)
}
\arguments{
\item{x}{CVlglasso object}

\item{xvar}{character which specify the x axis of the plot}

\item{...}{other plot arguments}
}
\value{
If \code{group} is NULL in \code{CVlglasso}, then a line plot will produced; otherwise, a heatmap will be produced.
}
\description{
Title
}
