% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{Simulate}
\alias{Simulate}
\title{Simulate longitudinal data from one-stage/two-stage model}
\usage{
Simulate(
  type = c("homo", "heter"),
  n,
  p,
  m1,
  m2,
  tt,
  tau,
  alpha = 1,
  group = 1
)
}
\arguments{
\item{type}{which type of data you are generating. There are two options. One is \code{homo} which generates subjects with identical
temporal correlation parameter. The other is \code{heter} which generates subjects with different temporal correlation parameter.}

\item{n}{the number of subjects in the data set}

\item{p}{the dimension of the normal distribution}

\item{m1}{the number of edges in true networks}

\item{m2}{the edge difference between two networks}

\item{tt}{the average time points for each subject}

\item{tau}{the true dampening rate in homogeneous models}

\item{alpha}{the true parameter in exponential distribution of tau when \code{type}
is \code{heter}}

\item{group}{a scalar of 1 or 2,  indicating one-stage or two-stage model.}
}
\value{
a data list. It include the data generated, true networks underlying the data,true tau.
If \code{type} is \code{heter}, then true parameter \code{alpha} is included as well.
}
\description{
This function randomly generates precision matrices (networks)
and then  simulates normal data that follows these network structures.
}
