% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSimulate2.R
\name{Simulate}
\alias{Simulate}
\title{Title}
\usage{
Simulate(
  type = c("general", "longihomo", "longiheter"),
  n = 20,
  p = 20,
  m1 = 20,
  m2 = 10,
  m3 = 3,
  cc = diag(m3),
  tau = c(2, 1),
  alpha = 2
)
}
\arguments{
\item{type}{what type of dat should be generated}

\item{n}{number of subjects}

\item{p}{number of nodes}

\item{m1}{number of edges}

\item{m2}{number of edges between general and individual network}

\item{m3}{number of correlated data within one subject}

\item{cc}{correlation matrix between tissues}

\item{tau}{parameters in correlation matrix}

\item{alpha}{parameter in correlation matrix}
}
\value{
data frame
}
\description{
Title
}
