% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionV1.R
\name{Simulate}
\alias{Simulate}
\title{Simulate data from a given network model}
\usage{
Simulate(
  type = c("general", "longihomo", "longiheter"),
  n = 20,
  p = 20,
  m1 = 20,
  m2 = 10,
  m3 = 3,
  tt = 5,
  cc = diag(m3),
  tau = c(2, 1),
  alpha = 2
)
}
\arguments{
\item{type}{what type of dat should be generated. There are three types of data that can be simulated through the function, which are
\code{general}, \code{longihomo} and \code{longiheter} respectively.}

\item{n}{number of subjects}

\item{p}{number of nodes}

\item{m1}{number of edges}

\item{m2}{number of edges between general and individual network (only for \code{longiheter})}

\item{m3}{number of correlated data within one subject}

\item{cc}{correlation matrix between tissues (only for \code{general})}

\item{tau}{parameters in correlation matrix (only for \code{longihomo})}

\item{alpha}{parameter in correlation matrix (only for \code{longiheter})}
}
\value{
list that include the relevant inputs and data generated.
}
\description{
This function generates three types of data based on the given network type. These data can then be used to test the
efficiency of algorithms.
}
\details{
\code{Simulate} function generate three types of data that can be used to test
the efficacy of  network identification algorithms.The first type \code{general} generates
data of which the covariance between tissues has no parametric  structure; the second type \code{longihomo} generates data which has
a same precision matrix that applies to  all subjects; the third type \code{longiheter} generates data in which subjects have
different precision matrix.
}
